refactor(cli): consolidate constants and update default limits across CLI commands
- Replaced references to deprecated constants in CLI commands with updated constants from the new location.
- Standardized the default limit for dataset and model listing commands to use DEFAULT_PAGE_SIZE for consistency.
- Cleaned up the constants file by removing unused definitions related to list limits and error messages.
- Adjusted related tests to reflect the changes in default limit handling.
