+++ osmosis_ai/platform/api/client.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING, Any, Literal
 from urllib.parse import quote, urlencode
 
 from osmosis_ai.platform.auth.platform_client import platform_request
+from osmosis_ai.platform.constants import DEFAULT_PAGE_SIZE
 
 from .models import (
     DatasetAffectedResources,
@@ -126,7 +127,7 @@ class OsmosisClient:
 
     def list_projects(
         self,
-        limit: int = 50,
+        limit: int = DEFAULT_PAGE_SIZE,
         offset: int = 0,
         *,
         credentials: Credentials | None = None,
@@ -244,7 +245,7 @@ class OsmosisClient:
 
     def list_datasets(
         self,
-        limit: int = 50,
+        limit: int = DEFAULT_PAGE_SIZE,
         offset: int = 0,
         *,
         credentials: Credentials | None = None,
@@ -296,7 +297,7 @@ class OsmosisClient:
     def list_training_runs(
         self,
         project_id: str,
-        limit: int = 20,
+        limit: int = DEFAULT_PAGE_SIZE,
         offset: int = 0,
         *,
         credentials: Credentials | None = None,
@@ -375,7 +376,7 @@ class OsmosisClient:
     def list_base_models(
         self,
         project_id: str,
-        limit: int = 50,
+        limit: int = DEFAULT_PAGE_SIZE,
         offset: int = 0,
         *,
         credentials: Credentials | None = None,
@@ -387,7 +388,7 @@ class OsmosisClient:
     def list_output_models(
         self,
         project_id: str,
-        limit: int = 50,
+        limit: int = DEFAULT_PAGE_SIZE,
         offset: int = 0,
         *,
         credentials: Credentials | None = None,
@@ -437,7 +438,7 @@ class OsmosisClient:
     def fetch_all_models(
         self,
         project_id: str,
-        limit: int = 50,
+        limit: int = DEFAULT_PAGE_SIZE,
         offset: int = 0,
         *,
         credentials: Credentials | None = None,
